feat: implement key management API and update related components
