+++ server/api/root.ts
@@ -1,4 +1,4 @@
-import { postRouter } from "./routers/post";
+import { keyRouter } from "./routers/key";
 import { createCallerFactory, createTRPCRouter } from "./trpc";
 
 
@@ -8,7 +8,7 @@ import { createCallerFactory, createTRPCRouter } from "./trpc";
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-  post: postRouter,
+  key: keyRouter,
 });
 
 // export type definition of API
